+++ src/main/java/com/anaplan/connector/AnaplanConnectorProperties.java
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package org.mule.modules.anaplan.connector;
+package com.anaplan.connector;
 
 
 import java.util.HashMap;
 
-import org.mule.modules.anaplan.connector.exceptions.ConnectorPropertiesException;
-import org.mule.modules.anaplan.connector.utils.LogUtil;
+import com.anaplan.connector.exceptions.ConnectorPropertiesException;
+import com.anaplan.connector.utils.LogUtil;
 
 
 /**
