+++ src/main/java/com/anaplan/connector/utils/AnaplanExecuteAction.java
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package org.mule.modules.anaplan.connector.utils;
+package com.anaplan.connector.utils;
 
-import org.mule.modules.anaplan.connector.AnaplanConnection;
-import org.mule.modules.anaplan.connector.AnaplanResponse;
-import org.mule.modules.anaplan.connector.exceptions.AnaplanOperationException;
 
 import com.anaplan.client.AnaplanAPIException;
+import com.anaplan.connector.AnaplanConnection;
+import com.anaplan.connector.AnaplanResponse;
+import com.anaplan.connector.exceptions.AnaplanOperationException;
 
 /**
  * Used to Delete data from an Anaplan model.
