+++ src/main/java/com/anaplan/connector/utils/AnaplanExportOperation.java
@@ -14,13 +14,10 @@
  * limitations under the License.
  */
 
-package org.mule.modules.anaplan.connector.utils;
+package com.anaplan.connector.utils;
 
 import java.io.IOException;
 
-import org.mule.modules.anaplan.connector.AnaplanConnection;
-import org.mule.modules.anaplan.connector.AnaplanResponse;
-import org.mule.modules.anaplan.connector.exceptions.AnaplanOperationException;
 
 import com.anaplan.client.AnaplanAPIException;
 import com.anaplan.client.Export;
@@ -28,6 +25,9 @@ import com.anaplan.client.Model;
 import com.anaplan.client.ServerFile;
 import com.anaplan.client.Task;
 import com.anaplan.client.TaskStatus;
+import com.anaplan.connector.AnaplanConnection;
+import com.anaplan.connector.AnaplanResponse;
+import com.anaplan.connector.exceptions.AnaplanOperationException;
 
 
 /**
