+++ src/main/java/com/anaplan/connector/utils/AnaplanImportOperation.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.mule.modules.anaplan.connector.utils;
+package com.anaplan.connector.utils;
 
 import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
@@ -24,9 +24,6 @@ import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.mule.modules.anaplan.connector.AnaplanConnection;
-import org.mule.modules.anaplan.connector.AnaplanResponse;
-import org.mule.modules.anaplan.connector.exceptions.AnaplanOperationException;
 
 import com.anaplan.client.AnaplanAPIException;
 import com.anaplan.client.CellWriter;
@@ -37,6 +34,9 @@ import com.anaplan.client.Task;
 import com.anaplan.client.TaskResult;
 import com.anaplan.client.TaskResultDetail;
 import com.anaplan.client.TaskStatus;
+import com.anaplan.connector.AnaplanConnection;
+import com.anaplan.connector.AnaplanResponse;
+import com.anaplan.connector.exceptions.AnaplanOperationException;
 import com.google.gson.JsonSyntaxException;
