+++ src/main/java/com/anaplan/connector/utils/BaseAnaplanOperation.java
@@ -14,11 +14,8 @@
  * limitations under the License.
  */
 
-package org.mule.modules.anaplan.connector.utils;
+package com.anaplan.connector.utils;
 
-import org.mule.modules.anaplan.connector.AnaplanConnection;
-import org.mule.modules.anaplan.connector.AnaplanResponse;
-import org.mule.modules.anaplan.connector.exceptions.AnaplanOperationException;
 
 import com.anaplan.client.Action;
 import com.anaplan.client.AnaplanAPIException;
@@ -27,6 +24,9 @@ import com.anaplan.client.Service;
 import com.anaplan.client.Task;
 import com.anaplan.client.TaskStatus;
 import com.anaplan.client.Workspace;
+import com.anaplan.connector.AnaplanConnection;
+import com.anaplan.connector.AnaplanResponse;
+import com.anaplan.connector.exceptions.AnaplanOperationException;
 
 
 /**
